+++ app/src/main/java/woowacourse/movie/main/model/Movie.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.model
+package woowacourse.movie.main.model
 
 import androidx.annotation.DrawableRes
