+++ app/src/main/java/woowacourse/movie/main/model/MovieRepository.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.model
+package woowacourse.movie.main.model
 
 import woowacourse.movie.R
