+++ app/src/main/res/layout/activity_movie_detail.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".view.MovieDetailActivity">
+    tools:context=".detail.view.MovieDetailActivity">
 
     <ScrollView
         android:layout_width="match_parent"
