+++ app/src/test/java/woowacourse/movie/main/model/MovieReservationCountTest.kt
@@ -1,7 +1,8 @@
-package woowacourse.movie.model
+package woowacourse.movie.main.model
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
+import woowacourse.movie.detail.model.MovieReservationCount
 
 class MovieReservationCountTest {
     @Test
